Revised function name and format
