+++ patterns/hexagonal-architecture-quickstart/src/main/java/com/baeldung/LibraryApplication.java
@@ -4,10 +4,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Import;
 
-import com.baeldung.rest.LibraryRestAdaptor;
+import com.baeldung.rest.LibraryRestController;
 
 @SpringBootApplication
-@Import(LibraryRestAdaptor.class)
+@Import(LibraryRestController.class)
 public class LibraryApplication {
 
 	public static void main(String[] args) {
