+++ patterns/hexagonal-architecture-quickstart/src/main/java/com/baeldung/exception/BorrowerNotFoundException.java
