+++ patterns/hexagonal-architecture-quickstart/src/main/java/com/baeldung/pojo/Book.java
