+++ patterns/hexagonal-architecture-quickstart/src/main/java/com/baeldung/repo/BookRepository.java
